Change globalRepos to withGlobalRepos
In other words, give it a callback, changing the type from

    globalRepos :: GlobalFlags -> [Repo]

to

    Verbosity -> GlobalFlags -> ([Repo] -> IO a) -> IO a

This will be necessary for repositories that need to do some
repository-specific initialization (even if we don't currently have any, we
will soon). The Verbosity flag is not used yet, but will be.

This just changes the type and deals with the fallout.