Extend repo config with security settings
This extends `RemoteRepo` with new fields

``` haskell
-- | Enable secure access to Hackage?
remoteRepoSecure :: Bool,

-- | Root key IDs (for bootstrapping)
remoteRepoRootKeys :: [String],

-- | Threshold for verification during bootstrapping
remoteRepoKeyThreshold :: Int,
```

and modifies the parser accordingly. This does not yet require a dependency on
the security library.

NOTE: We suggest to make the security functionality opt-in, rather than
opt-out, for the first official release. Thus, the 'secure' field will be set
to 'False' when initializing a `~/.cabal/config` file, and existing
`~/.cabal/config` files will be interpreted as if they had `secure: False`.