Extend the backwards compat Setup.hs flags filter
Cabal < 1.12.0 doesn't know about '--enable/disable-executable-dynamic'
or '--enable/disable-library-coverage'.

This is a problem in general, if rare, but it shows up in the
nix-local-build branch more obviously since it always passes all flags,
including defaults (it does this because the defaults should be
determined by the current version, not the older version).