Testing with GHC 7.0 and 7.2.
In the process, I refactored the test suite a bit to be more robust
and handle the --with-compiler more correctly in most cases.

One note: I turned off the HPC tests with the GHC and WITH_GHC
parameters don't match.  In principle this should work but there
is some sort of bug.  I don't plan on fixing the bug.

Signed-off-by: Edward Z. Yang <ezyang@cs.stanford.edu>