FileMonitor: reuse known file hashes when updating
We know the previous mtime and content hash for all the files in the
previous file set, so reuse these content hashes when we determine they
are still valid (which is when the mtime is the same). This saves
re-hashing files that have not changed.

No test for this as such since it's a performance rather than
correctness thing, but checked ad-hoc that the cache does get a perfect
hit rate.

Fixes one of two parts of issue #3019.