Turn ReadyPackage into a newtype wrapper.
Previously, ReadyPackage was a ConfiguredPackage elaborated with
a dependencies data structure which had InstalledPackageInfo
rather than ConfiguredId. Well, it turned out that we only
used the data from ConfiguredId! So that extra info is useless.

Instead, ReadyPackage is now purely a newtype wrapper for type
safety; a reminder that not all ConfiguredPackages can be built,
only the ready ones.

Signed-off-by: Edward Z. Yang <ezyang@cs.stanford.edu>