Pass per-package program config through
That is, the --with-$PROG program locations, the --$PROG-option(s)
program arguments and the --extra-prog-path search paths. These can be
passed on the command line or specified in the cabal.project file.

In the project file they can be specified for all local package or on a
per-package basis.

The compiler and its related programs are handled specially. Their
location cannot be specified independently for each package. A
subsequent patch will add the checks for this. Otherwise we just pass
the program config though.

Program options specified on the command line currently apply to all
local packages, not just the targets of the command. We may want to
revisit this.