Add the current goal to the initial conflict set while backjumping.
This is relevant for unknown packages. Without the current
goal in the initial conflict set, the unknown package itself
does not end up in its own conflict set.

But we want it to, if only to make sure that the error slicing
machinery keeps the message about the unknown goal around.

It is also the correct thing to do. The initial conflict set
corresponds to a virtual "avoid" choice underneath the package
goal. That choice corresponds to avoiding assigning any proper
value to the goal at all, which might be possible if the goal
was not needed. The conflict set for this "avoid" goal is always
between the current package and its immediate goal reason.

I have added comments to the code explaining this as well.