Replace Goal with Var where possible; remove Unknown goal reason.
In several places where we used to work with Goals mainly to maintain
the goal reasons for constructing conflict sets, we can now do so
without, and just use Vars instead.

This also has the advantage that when creating dependencies during
index conversion, we now do not have to invent goal reasons. We used
to do that by having an "Unknown" goal reason. This is now no longer
necessary.