Fix bugs with bash completion
The bash completion suffers from a number of issues,
some fix with this patch;

With this patch we expand the path to cabal as well as its arguments.
For example;
~/code/cabal configure --with-ghc=$GHC <tab><tab>
will now work. Both ~ and the $GHC variable will be expanded.

If there are no .cabal files, or multiple files, the script will no longer
crash.

We still invoke cabal to get completions. If cabal detects an error and writes
errors to STDERR, these will no longer be printed to the console. Instead we
don't offer any completions.