Add integration tests for setup script handling
Covers 3 of the 4 possible cases:
1. explicit custom setup deps
2. custom setup with implicit/default deps
4. non-custom setup using the internal cabal lib version

case 3 is a non-custom setup but where we're forced to use an external
cabal lib version. This case is hard to test since it only happens when
it's a newer (not older) Cabal lib version that the package requires,
e.g. a .cabal file that specifies cabal-version: >= 2.0.

Also, add a --with-ghc option to the integration test suite, which lets us
more easily test with different ghc versions.

Also, don't use parallel builds in any of the integration tests, as the
self-exec method will not work, and some tests need to install deps for
some ghc versions.