haddock/hscolour: fix highlighted source location
When generating haddocks with the `--for-hackage` switch, the generated
haddocks are placed in a different directory than the normal ones, which
includes the package id instead of just the package name. When we ran
hscolour, we didn't respect this, so the highlighted source would not
be placed in the correct directory and thus was missing from the tarball.
This patch fixes that.

Fixes #3451