executeInstallPlan: mode to stop as soon as possible after failure
Add a keepGoing :: Bool arg to control whether to keep going after a
build failure or whether to try and stop as soon as possible.

This makes use of the new JobControl interface to try cancelling
outstanding jobs and also for checking if there are jobs in progress.

For now the call sites are hard coded to use keepGoing = False.