Make a copy of InstallPlan named SolverInstallPlan.
Now we copy-paste the contents of InstallPlan into SolverInstallPlan,
thus giving it a separate set of types.  For now, we don't do anything
else, e.g., remove unnecessary functions or specialize.

We need a new function 'fromSolverInstallPlan' akin to
'mapPreservingGraph' which can take an InstallPlan from the
old solver install plan to the new one.

Signed-off-by: Edward Z. Yang <ezyang@cs.stanford.edu>